now only portal admins can upload files
we'll have to clarify some things for the upload/download
we have an admin and a portal admin: who can do what?
would it be a good idea to split the document controller to upload and download controllers? (if upload is restricted, the checks for could the access checks be done for the whole controller)